requirements.txt
fastapi
python-multipart
"uvicorn[standard]"
python-jose